Add second goss check method
Add a second method to check the goss results by not running the goss
test expliclty but check the output of the goss test that was run by an
external program (i.e. for prometheus).

Add 2 variables to control which behavior is used and where the goss output file is located